pytest/conftest fix, discovering this feature
