+++ tests/conftest.py
@@ -1,0 +1,50 @@
+# Copyright (c) Facebook, Inc. and its affiliates. All rights reserved.
+#
+# This source code is licensed under the BSD license found in the
+# LICENSE file in the root directory of this source tree.
+
+# Copyright 2019 Kakao Brain
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import functools
+from typing import Callable
+
+import pytest
+import torch
+
+
+@pytest.fixture(autouse=True)
+def manual_seed_zero() -> None:
+    torch.manual_seed(0)
+
+
+def cuda_sleep_impl(seconds: int, cycles_per_ms: int) -> None:
+    torch.cuda._sleep(int(seconds * cycles_per_ms * 1000))
+
+
+@pytest.fixture(scope="session")
+def cuda_sleep() -> Callable:
+    # Warm-up CUDA.
+    torch.empty(1, device="cuda")
+
+    # From test/test_cuda.py in PyTorch.
+    start = torch.cuda.Event(enable_timing=True)
+    end = torch.cuda.Event(enable_timing=True)
+    start.record()
+    torch.cuda._sleep(1000000)
+    end.record()
+    end.synchronize()
+    cycles_per_ms = 1000000 / start.elapsed_time(end)
+
+    return functools.partial(cuda_sleep_impl, cycles_per_ms=cycles_per_ms)
